Support Qwen3-4B training: GQA LoRA fix, bf16 weight sync, logprob chunking
- Fix dummy LoRA dimensions for GQA models (q_proj out != hidden_size)
  that caused vLLM server hang with Qwen3-4B
- Allow bf16 model weights with vLLM by converting to float32 only
  during weight serialization (.float().numpy())
- Add --logprob_batch_size to chunk logprob forward pass for large-vocab
  models (Qwen3 151k vocab causes OOM without chunking)
- Fix sweep.py: strip retain_pass params from baselines, pin GPU in
  non-MPS mode, detect external vLLM server to skip per-GPU launch
- Add qwen4b_leetcode_routing sweep config

+++ vllm_client.py
@@ -44,7 +44,7 @@ def _extract_weights_from_model(model):
                 attr = getattr(module, key, None)
                 if attr is not None:
                     t = attr.weight.data
-                    layer_data[key] = t.detach().cpu().numpy().tobytes()
+                    layer_data[key] = t.detach().cpu().float().numpy().tobytes()
                     if key not in shapes:
                         shapes[key] = list(t.shape)
                 else:
